Made new scene
scene includes: npc object. the npc object has a script and a collider attached to to. the collider is the size of a grid square .

dialogue will be shown when the player is in the adjacent square (yet to be done)

dialogue will be advanced if the player presses the interact key while in said square (yet to be done)