ColorBlindness Settings
Made a dropdown for various forms of colorblindness. Saves using PlayerPrefs.

Colorblindness is hard to quantify, and there is no one correct solution to this. By allowing so many modes, it allows users to find one best for them.

Since there is no settings menu yet, applying these effects on a per-scene basis will be a future project.

Currently used in GYM_ColorSettings