OPTIONS MENU!!!!!!
It has various options:
Resolution, fullscreen toggle, audio sliders, color filter.

It also includes tooltips and subtitles toggles, but they are not function yet.

All of the settings are saved using player prefs.

Other cool features:
- 4 tabs (DIsplay, Audio, Accessibility, Gameplay)
- cursor changes when hovering over buttons
- the tabs work and show active tabs

All located in the GYM_OptionsMenu scene

Loading certain settings is not operational yet, however this is just to get the actual menu in. Loading settings will be in a separate commit.