NPCRoom Checks
NPC dialogue must be finished for the doors to open. This is saved using PlayerPrefs. Doors need to be manually assigned in the NPC's NPCScript.

Will work on an infinite number of scenes. Only works if there is one NPC in the room. Will be worked on further if requested.