Changed the canvas for the Level Select to accommodate automated buttons
Everything I've done is directly added and saved to the Base_LevelPrefab so keep that in mind.
I've changed the orientation of the Level Select panel. I've gotten rid of the title and click here button (the debug manager still holds the old buttons though they aren't visible because I got rid of the click here button). So far everything is labeled with Level and their index number. The buttons work, but I'd like  help to make sure I didn't mess something else up getting it up and running.