Added a crude way to change the sprites of the options menu buttons
I'm not completely sure if it works cause I need alternate images to be put into the SerializeFields. Also I might need to come up with a way to get the initial image to show up when it gets clicked, but I'll know that once everything else can be tested.