Changed the way ChangeImage works
Now ChangeImage goes through everything in the _ogSprites dict and changes the image based on what isn't clicked. I also changed HighlightButton so the UI is inline with what Rider wanted it to look like again. Only thing is the highlight bit which if we want it in there makes it necessary to have the tab as a parameter.