Mirror Mother and Dialogue Sounds
Added sound effects for Mirror Mother walking and dialogue boxes popping up and clicking through the text. Dialogue has been set up to allow a simple switch between the two sound effects. The first (click) is for opening and closing the dialogue box and the second one (prompt) is for going from one box to another.