Adds position check for dying to mirror and copy enemies
Adds in the position check for killing the player from the normal enemy behavior script because the current check can fail on special tiles like moveable walls. Easiest way to check this bug fix is in P_DuelPadSwMwSpLpHb at the top left moveable wall since before this fix stepping on the same moveable wall tile with a mirror enemy didn't kill the player.