Puzzle Iterations for BlitzWrap, SquareFreeze, and DuelPad
Puzzle Iterations for BlitzWrap: I reduced the size of the map to reduce unnecessary or unused tiles

SquareFreeze: I took away two tiles near the door so the player has to unfreeze all enemies to leave.

DuelPad: I reduced the size of the map so the player is waiting less often for enemies to walk by. I also adjusted the placement of the moving wall in the center so it is not obstructing the Enemy's pathing. I also reduced the play area for the mirror enemy to reduce confusion.