First Button Focuses and Unfocuses as needed
The bug is no more and the first button will now only be highlighted if a controller is in used. If a keyboard was used to pause the game a mouse is assumed to be used for menuing. Also got a rid of unnecessary code that didn't need to be used due to the change.